Move menu into own folder and more work on stubbing
