+++ src/client/components/chat/chat.scss
@@ -12,207 +12,6 @@
   flex-direction: row;
 }
 
-.chat-menu {
-  position: relative;
-  height: 100vh;
-  z-index: 101;
-  border-right: 1px solid #f7f6f6;
-  width: 75px;
-
-  .chat-menu-container {
-    height: 100%;
-    position: relative;
-    z-index: 3;
-  }
-
-  .minibar {
-    position: relative;
-    z-index: 2;
-    height: 100%;
-    overflow: hidden;
-    border-right: 1px solid #f1f1f1;
-
-    &::before {
-      content: ' ';
-      display: inline-block;
-      position: absolute;
-      top: 0;
-      left: 0;
-      width: 100%;
-      height: 100%;
-      opacity: 0.3;
-      transform: translate3d(0, 0, 0);
-    }
-
-    ul {
-      position: absolute;
-      // overflow-x: hidden;
-      overflow: hidden;
-      top: 0;
-      right: 0;
-      bottom: 0;
-      left: 0;
-      display: flex;
-      flex-direction: column;
-      margin-left: 0;
-      margin-bottom: 0;
-      list-style: none;
-    }
-
-    li {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      text-align: center;
-      width: 75px;
-      min-width: 75px;
-      height: 65px;
-      min-height: 65px;
-      line-height: 65px;
-      position: fixed;
-      cursor: default;
-    }
-
-    button {
-      display: block;
-      width: 100%;
-      height: 100%;
-      padding: 0;
-      background-color: transparent;
-      border: 0;
-      transition: opacity 0.1s ease-in;
-      color: inherit;
-      text-decoration: none;
-      line-height: 0;
-      text-align: center;
-      opacity: 1;
-      outline-color: $razz;
-      outline-style: none;
-      outline-width: 0;
-      vertical-align: middle;
-
-      &:focus {
-        outline: none;
-      }
-    }
-  }
-
-  .minibar-inner {
-    position: absolute;
-    top: 325px;
-    bottom: 130px;
-    height: auto;
-    width: 100%;
-    overflow: hidden;
-  }
-
-  .minibar-convos {
-    background-color: $violet;
-    color: $london;
-    top: 0;
-    left: 0;
-
-    &::after {
-      content: '';
-      display: block;
-      width: 2px;
-      height: 100%;
-      position: absolute;
-      top: 0;
-      right: -1px;
-      background-color: $violet;
-    }
-  }
-
-  .minibar-search {
-    top: 65px;
-    left: 0;
-    color: $maroon;
-  }
-}
-
-.chat-menu-panel {
-  box-sizing: border-box;
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  padding-left: 100%;
-  border-right: 1px solid #f1f1f1;
-  background-color: #fdfdfd;
-  transition: transform 0.1s linear 0.05s;
-  transform: translateX(-100%);
-  z-index: 1;
-
-  &.active {
-    transform: translateX(0);
-  }
-
-  .brand-container {
-    position: relative;
-    display: flex;
-    align-items: center;
-    height: 65px;
-    background-color: $violet;
-    padding-left: 2.5rem;
-    color: $london;
-
-    &::after {
-      content: '';
-      display: block;
-      width: 2px;
-      height: 100%;
-      position: absolute;
-      top: 0;
-      right: -1px;
-      background-color: $violet;
-    }
-  }
-
-  .panel-header {
-    height: 65px;
-    position: relative;
-    min-height: 65px;
-    overflow: hidden;
-    text-align: left;
-
-    .panel-header-convos {
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 1.3rem;
-      margin-bottom: 0;
-      // transform: translateY(-100%);
-      // opacity: 0;
-      transition: transform 0.2s ease-in-out, opacity 0.2s ease-out, background-color 0.2s ease;
-      box-sizing: border-box;
-      display: flex;
-      align-items: center;
-      // transition: background-color 0.2s ease;
-
-      h2 {
-        text-rendering: optimizeLegibility;
-        -webkit-font-smoothing: antialiased;
-        max-width: 90%;
-        margin: 0;
-        overflow: hidden;
-        font-weight: 400;
-        font-size: 15px;
-        white-space: nowrap;
-        text-overflow: ellipsis;
-        text-transform: uppercase;
-        padding: 0 0 0 2.6rem;
-        line-height: 65px;
-        color: #a2a3a3;
-      }
-    }
-  }
-
-  .panel-inner {
-    width: 264px;
-  }
-}
-
 .chat-header-wrapper {
   width: calc(100% - 75px);
   height: 100%;
