+++ src/client/components/chat/menu/ChatMenu.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import            './menu.scss';
 
 export class ChatMenu extends React.Component {
   constructor(props) {
@@ -17,30 +18,32 @@ export class ChatMenu extends React.Component {
   handleClick() {
     this.toggleActive();
   }
-  handleMouseLeave() {
+  handleMouseLeave(e) {
+    if (e.clientX < 74) return; // ignore leaving to minibar
     this.toggleActive();
   }
   render() {
     const active = this.state.active ? 'active' : '';
     return (
-      <aside
-        className="chat-menu"
-        onMouseLeave={this.handleMouseLeave}
-      >
+      <aside className="chat-menu">
         <section className="chat-menu-container">
           <nav className="minibar">
             <div className="minibar-inner">
               <ul>
                 <li className="minibar-convos">G</li>
                 <li className="minibar-search">
-                  <button onClick={this.handleClick}>
+                  <button className="minibar-button" onClick={this.handleClick}>
                     Q
                   </button>
                 </li>
               </ul>
             </div>
           </nav>
-          <section className={'chat-menu-panel ' + active}>
+          <section
+            // className="chat-menu-panel active"
+            className={'chat-menu-panel ' + active}
+            onMouseLeave={this.handleMouseLeave}
+          >
             <div className="brand-container">Gitter Clone</div>
             <header className="panel-header">
               <div className="panel-header-convos">
@@ -48,7 +51,16 @@ export class ChatMenu extends React.Component {
               </div>
             </header>
             <div className="panel-inner">
-              {/* links to Conversations */}
+              <footer className="panel-footer">
+                <div className="panel-footer-convos">
+                  <a className="create-rooom-button">Add a room</a>
+                </div>
+              </footer>
+              <div className="panel-content scroller">
+                <section className="primary-collection">
+                  {/* links to Conversations */}
+                </section>
+              </div>
             </div>
           </section>
         </section>
